chore: renamed about folder to settings due to its only child
