+++ lib/ui/map/map_overlay/map_settings_sheet.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:munich_ways/ui/about/settings_sheet_content.dart';
+import 'package:munich_ways/ui/settings/settings_sheet_content.dart';
 import 'package:munich_ways/ui/map/map_screen_model.dart';
 import 'package:munich_ways/ui/widgets/bottom_sheet.dart';
